refactor(level): Rename `map_limit` mod to `limit`
